+++ common/src/main/java/io/github/flemmli97/flan/integration/currency/CommandCurrency.java
@@ -1,19 +1,20 @@
 package io.github.flemmli97.flan.integration.currency;
 
-import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import dev.architectury.injectables.annotations.ExpectPlatform;
-import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
+
+import java.util.function.Consumer;
 
 public class CommandCurrency {
 
     @ExpectPlatform
-    public static int sellClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
+    public static boolean sellClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
         throw new AssertionError();
     }
 
     @ExpectPlatform
-    public static int buyClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
+    public static boolean buyClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
         throw new AssertionError();
     }
 }