+++ fabric/src/main/java/io/github/flemmli97/flan/integration/currency/fabric/CommandCurrencyImpl.java
@@ -1,9 +1,5 @@
 package io.github.flemmli97.flan.integration.currency.fabric;
 
-import com.mojang.brigadier.Command;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
-import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import io.github.flemmli97.flan.Flan;
 import io.github.flemmli97.flan.claim.PermHelper;
 import io.github.flemmli97.flan.config.ConfigHandler;
@@ -11,57 +7,57 @@ import io.github.flemmli97.flan.player.PlayerClaimData;
 import io.github.gunpowder.entities.StoredBalance;
 import io.github.gunpowder.modelhandlers.BalanceHandler;
 import net.minecraft.ChatFormatting;
-import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
 
 import java.math.BigDecimal;
+import java.util.function.Consumer;
 
 public class CommandCurrencyImpl {
 
-    public static int sellClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
-        if (!Flan.gunpowder) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED), false);
-            return 0;
+    public static boolean sellClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
+        if (value == -1) {
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellDisabled"), ChatFormatting.DARK_RED));
+            return false;
         }
-        if (ConfigHandler.config.sellPrice == -1) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellDisabled"), ChatFormatting.DARK_RED), false);
-            return 0;
+        if (Flan.gunpowder) {
+            PlayerClaimData data = PlayerClaimData.get(player);
+            if (data.getAdditionalClaims() - Math.max(0, data.usedClaimBlocks() - data.getClaimBlocks()) < blocks) {
+                message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellFail"), ChatFormatting.DARK_RED));
+                return false;
+            }
+            StoredBalance bal = BalanceHandler.INSTANCE.getUser(player.getUUID());
+            BigDecimal price = BigDecimal.valueOf(blocks * value);
+            bal.setBalance(bal.getBalance().add(price));
+            BalanceHandler.INSTANCE.updateUser(bal);
+            data.setAdditionalClaims(data.getAdditionalClaims() - blocks);
+            message.accept(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("sellSuccess"), blocks, price), ChatFormatting.GOLD));
+            return true;
         }
-        int amount = Math.max(0, IntegerArgumentType.getInteger(context, "amount"));
-        PlayerClaimData data = PlayerClaimData.get(context.getSource().getPlayerOrException());
-        if (data.getAdditionalClaims() - Math.max(0, data.usedClaimBlocks() - data.getClaimBlocks()) < amount) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellFail"), ChatFormatting.DARK_RED), false);
-            return 0;
-        }
-        StoredBalance bal = BalanceHandler.INSTANCE.getUser(context.getSource().getPlayerOrException().getUUID());
-        BigDecimal price = BigDecimal.valueOf(amount * ConfigHandler.config.sellPrice);
-        bal.setBalance(bal.getBalance().add(price));
-        BalanceHandler.INSTANCE.updateUser(bal);
-        data.setAdditionalClaims(data.getAdditionalClaims() - amount);
-        context.getSource().sendSuccess(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("sellSuccess"), amount, price), ChatFormatting.GOLD), false);
-        return Command.SINGLE_SUCCESS;
+        message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED));
+        return false;
     }
 
-    public static int buyClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
-        if (!Flan.gunpowder) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED), false);
-            return 0;
+    public static boolean buyClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
+        if (value == -1) {
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyDisabled"), ChatFormatting.DARK_RED));
+            return false;
         }
-        if (ConfigHandler.config.buyPrice == -1) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyDisabled"), ChatFormatting.DARK_RED), false);
-            return 0;
+        if (Flan.gunpowder) {
+            StoredBalance bal = BalanceHandler.INSTANCE.getUser(player.getUUID());
+            BigDecimal price = BigDecimal.valueOf(Math.max(0, blocks * value));
+            if (bal.getBalance().compareTo(price) >= 0) {
+                PlayerClaimData data = PlayerClaimData.get(player);
+                data.setAdditionalClaims(data.getAdditionalClaims() + blocks);
+                bal.setBalance(bal.getBalance().subtract(price));
+                BalanceHandler.INSTANCE.updateUser(bal);
+                message.accept(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("buySuccess"), blocks, price), ChatFormatting.GOLD));
+                return true;
+            }
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyFail"), ChatFormatting.DARK_RED));
+            return false;
         }
-        StoredBalance bal = BalanceHandler.INSTANCE.getUser(context.getSource().getPlayerOrException().getUUID());
-        int amount = Math.max(0, IntegerArgumentType.getInteger(context, "amount"));
-        BigDecimal price = BigDecimal.valueOf(amount * ConfigHandler.config.buyPrice);
-        if (bal.getBalance().compareTo(price) >= 0) {
-            PlayerClaimData data = PlayerClaimData.get(context.getSource().getPlayerOrException());
-            data.setAdditionalClaims(data.getAdditionalClaims() + amount);
-            bal.setBalance(bal.getBalance().subtract(price));
-            BalanceHandler.INSTANCE.updateUser(bal);
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("buySuccess"), amount, price), ChatFormatting.GOLD), false);
-            return Command.SINGLE_SUCCESS;
-        }
-        context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyFail"), ChatFormatting.DARK_RED), false);
-        return 0;
+        message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED));
+        return false;
     }
 }