+++ forge/src/main/java/io/github/flemmli97/flan/integration/currency/forge/CommandCurrencyImpl.java
@@ -1,9 +1,5 @@
 package io.github.flemmli97.flan.integration.currency.forge;
 
-import com.mojang.brigadier.Command;
-import com.mojang.brigadier.arguments.IntegerArgumentType;
-import com.mojang.brigadier.context.CommandContext;
-import com.mojang.brigadier.exceptions.CommandSyntaxException;
 import dicemc.money.MoneyMod;
 import dicemc.money.storage.MoneyWSD;
 import io.github.flemmli97.flan.Flan;
@@ -11,56 +7,56 @@ import io.github.flemmli97.flan.claim.PermHelper;
 import io.github.flemmli97.flan.config.ConfigHandler;
 import io.github.flemmli97.flan.player.PlayerClaimData;
 import net.minecraft.ChatFormatting;
-import net.minecraft.commands.CommandSourceStack;
+import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerPlayer;
 
 import java.util.UUID;
+import java.util.function.Consumer;
 
 public class CommandCurrencyImpl {
 
-    public static int sellClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
-        if (!Flan.diceMCMoneySign) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED), false);
-            return 0;
+    public static boolean sellClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
+        if (value == -1) {
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellDisabled"), ChatFormatting.DARK_RED));
+            return false;
         }
-        if (ConfigHandler.config.sellPrice == -1) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellDisabled"), ChatFormatting.DARK_RED), false);
-            return 0;
+        if (Flan.diceMCMoneySign) {
+            PlayerClaimData data = PlayerClaimData.get(player);
+            if (data.getAdditionalClaims() - Math.max(0, data.usedClaimBlocks() - data.getClaimBlocks()) < blocks) {
+                message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellFail"), ChatFormatting.DARK_RED));
+                return false;
+            }
+            double price = blocks * value;
+            MoneyWSD.get(player.getLevel()).changeBalance(MoneyMod.AcctTypes.PLAYER.key, player.getUUID(), price);
+            data.setAdditionalClaims(data.getAdditionalClaims() - blocks);
+            message.accept(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("sellSuccess"), blocks, price), ChatFormatting.GOLD));
+            return true;
         }
-        int amount = Math.max(0, IntegerArgumentType.getInteger(context, "amount"));
-        PlayerClaimData data = PlayerClaimData.get(context.getSource().getPlayerOrException());
-        if (data.getAdditionalClaims() - Math.max(0, data.usedClaimBlocks() - data.getClaimBlocks()) < amount) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("sellFail"), ChatFormatting.DARK_RED), false);
-            return 0;
-        }
-        double price = amount * ConfigHandler.config.sellPrice;
-        MoneyWSD.get(context.getSource().getLevel()).changeBalance(MoneyMod.AcctTypes.PLAYER.key, context.getSource().getPlayerOrException().getUUID(), price);
-        data.setAdditionalClaims(data.getAdditionalClaims() - amount);
-        context.getSource().sendSuccess(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("sellSuccess"), amount, price), ChatFormatting.GOLD), false);
-        return Command.SINGLE_SUCCESS;
+        message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED));
+        return false;
     }
 
-    public static int buyClaimBlocks(CommandContext<CommandSourceStack> context) throws CommandSyntaxException {
-        if (!Flan.diceMCMoneySign) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED), false);
-            return 0;
+    public static boolean buyClaimBlocks(ServerPlayer player, int blocks, float value, Consumer<Component> message) {
+        if (value == -1) {
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyDisabled"), ChatFormatting.DARK_RED));
+            return false;
         }
-        if (ConfigHandler.config.buyPrice == -1) {
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyDisabled"), ChatFormatting.DARK_RED), false);
-            return 0;
+        if (Flan.diceMCMoneySign) {
+            UUID uuid = player.getUUID();
+            MoneyWSD manager = MoneyWSD.get(player.getLevel());
+            double bal = manager.getBalance(MoneyMod.AcctTypes.PLAYER.key, uuid);
+            double price = blocks * value;
+            if (bal >= price) {
+                PlayerClaimData data = PlayerClaimData.get(player);
+                data.setAdditionalClaims(data.getAdditionalClaims() + blocks);
+                manager.changeBalance(MoneyMod.AcctTypes.PLAYER.key, uuid, -price);
+                message.accept(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("buySuccess"), blocks, price), ChatFormatting.GOLD));
+                return true;
+            }
+            message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyFail"), ChatFormatting.DARK_RED));
+            return false;
         }
-        UUID uuid = context.getSource().getPlayerOrException().getUUID();
-        MoneyWSD manager = MoneyWSD.get(context.getSource().getLevel());
-        double bal = manager.getBalance(MoneyMod.AcctTypes.PLAYER.key, uuid);
-        int amount = Math.max(0, IntegerArgumentType.getInteger(context, "amount"));
-        double price = amount * ConfigHandler.config.buyPrice;
-        if (bal >= price) {
-            PlayerClaimData data = PlayerClaimData.get(context.getSource().getPlayerOrException());
-            data.setAdditionalClaims(data.getAdditionalClaims() + amount);
-            manager.changeBalance(MoneyMod.AcctTypes.PLAYER.key, uuid, -price);
-            context.getSource().sendSuccess(PermHelper.simpleColoredText(String.format(ConfigHandler.langManager.get("buySuccess"), amount, price), ChatFormatting.GOLD), false);
-            return Command.SINGLE_SUCCESS;
-        }
-        context.getSource().sendSuccess(PermHelper.simpleColoredText(ConfigHandler.langManager.get("buyFail"), ChatFormatting.DARK_RED), false);
-        return 0;
+        message.accept(PermHelper.simpleColoredText(ConfigHandler.langManager.get("currencyMissing"), ChatFormatting.DARK_RED));
+        return false;
     }
 }